create ai endpoint for adding new record & renaming for better organization
