+++ app/Board.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Board extends Model
+{
+    protected $fillable = ['title, user_id'];
+}
