+++ app/Http/Controllers/Auth/VerificationController.php
@@ -25,7 +25,7 @@ class VerificationController extends Controller
      *
      * @var string
      */
-    protected $redirectTo = '/home';
+    protected $redirectTo = '/board';
 
     /**
      * Create a new controller instance.
