+++ app/Http/Controllers/BoardsController.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use App\Board;
+use Response;
+
+class BoardsController extends Controller
+{
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
+    /**
+     * Show the application dashboard.
+     *
+     * @return \Illuminate\Contracts\Support\Renderable
+     */
+    public function index()
+    {
+        $boards = Board::all();
+        return view('board', compact('boards'));
+    }
+
+    public function store(){
+
+        request()->validate(['title' => ['required','min:3']]);
+        $boards = new Board;
+        $boards->title = request('title');
+        $boards->user_id = auth()->id();
+        $boards->save();
+
+        return response()->json($boards);
+
+        // return redirect('/board');
+    }
+}
+
+
+
+
+
+
+
