+++ routes/web.php
@@ -17,4 +17,5 @@ Route::get('/', function () {
 
 Auth::routes();
 
-Route::get('/home', 'HomeController@index')->name('home');
+Route::get('/board', 'BoardsController@index');
+Route::post('/board', 'BoardsController@store');
